Add preCharge func to controllers & cleaned server
